Handle user config
