+++ clients/ui/frontend/src/shared/types.ts
@@ -3,6 +3,7 @@ import { ValueOf } from '~/shared/typeHelpers';
 //  TODO: [Data Flow] Get the status config params
 export type UserSettings = {
   username: string;
+  isAdmin?: boolean;
 };
 
 // TODO: [Data Flow] Add more config parameters
